incorporate and test different job flows
