+++ app/templates/index.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block content %}
-<h2>Harvest Actionszzz</h2>
+<h2>Harvest Actions</h2>
 
 <ul class="list-unstyled">
     {% if session['user'] %}
