+++ harvester/lib/load_manager.py
@@ -55,13 +55,13 @@ class LoadManager:
 
         logger.info("Load Manager :: Updated Harvest Jobs")
 
-    def start_job(self, job_id):
+    def start_job(self, job_id, job_type="harvest"):
         """task manager start interface,
         takes a job_id"""
         task_contract = {
             "app_guuid": HARVEST_RUNNER_APP_GUID,
-            "command": f"python harvester/harvest.py {job_id}",
-            "task_id": f"harvest-job-{job_id}",
+            "command": f"python harvester/harvest.py {job_id} {job_type}",
+            "task_id": f"harvest-job-{job_id}-{job_type}",
         }
 
         self.handler.start_task(**task_contract)
@@ -70,14 +70,14 @@ class LoadManager:
         logger.info(message)
         return message
 
-    def stop_job(self, job_id):
+    def stop_job(self, job_id, job_type="harvest"):
         """task manager stop interface,
         takes a job_id"""
         tasks = self.handler.get_all_app_tasks(HARVEST_RUNNER_APP_GUID)
         job_task = [
             (t["guid"], t["state"])
             for t in tasks
-            if t["name"] == f"harvest-job-{job_id}"
+            if t["name"] == f"harvest-job-{job_id}-{job_type}"
         ]
 
         if len(job_task) == 0:
