+++ tests/integration/app/fixtures/task_manager_all_tasks.json
@@ -4,8 +4,8 @@
         "created_at": "2024-10-18T16:35:58Z",
         "updated_at": "2024-10-18T16:36:33Z",
         "sequence_id": 198,
-        "name": "harvest-job-b22a8e9b-4906-4515-97ab-3a24b55a02b0",
-        "command": "python harvester/harvest.py 47442c62-716d-4678-947c-61990106685f",
+        "name": "harvest-job-b22a8e9b-4906-4515-97ab-3a24b55a02b0-harvest",
+        "command": "python harvester/harvest.py 47442c62-716d-4678-947c-61990106685f harvest",
         "state": "SUCCEEDED",
         "memory_in_mb": 1536,
         "disk_in_mb": 4096,
